+++ src/my-vue-app/src/utils/mitt.ts
@@ -1,6 +1,0 @@
-import mitt, { Emitter } from 'mitt';
-
-// 类型
-const emitter: Emitter<MittType> = mitt<MittType>();
-
-export default emitter;