+++ src/starshine-admin-web/src/api-services/index.ts
@@ -1,7 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-
-export * from "./api";
-export * from "./configuration";
-export * from "./models";
-