+++ src/starshine-admin-web/src/api-services/models/base/index.ts
@@ -1,4 +1,0 @@
-export * from './base-id-input';
-export * from './base-page-input';
-export * from './admin-result';
-export * from './paed-list-result';