+++ src/starshine-admin-web/src/api-services/models/logs/index.ts
@@ -1,5 +1,0 @@
-export * from './sys-log-vis';
-export * from './sys-log-op';
-export * from './sys-log-diff';
-export * from './sys-log-ex';
-export * from './sys-log';