+++ src/starshine-admin-web/src/layout/logo/index.vue
@@ -1,6 +1,6 @@
 <template>
 	<div class="layout-logo" v-if="setShowLogo" @click="onThemeConfigChange">
-		<img :src="logoMini" class="layout-logo-medium-img" v-if="showLogo" />
+		<img :src="logoMini" class="layout-logo-medium-img" />
 		<span>{{ themeConfig.globalTitle }}</span>
 	</div>
 	<div class="layout-logo-size" v-else @click="onThemeConfigChange">
@@ -22,10 +22,6 @@ const { themeConfig } = storeToRefs(storesThemeConfig);
 const setShowLogo = computed(() => {
 	let { isCollapse, layout } = themeConfig.value;
 	return !isCollapse || layout === 'classic' || document.body.clientWidth < 1000;
-});
-const showLogo = computed(() => {
-	let { layout } = themeConfig.value;
-	return layout !== 'columns';
 });
 // logo 点击实现菜单展开/收起
 const onThemeConfigChange = () => {
@@ -49,8 +45,6 @@ const onThemeConfigChange = () => {
 	span {
 		white-space: nowrap;
 		display: inline-block;
-		font-weight: 800;
-		font-size: 18px;
 	}
 	&:hover {
 		span {