+++ src/starshine-admin-web/src/stores/themeConfig.ts
@@ -18,7 +18,7 @@ export const useThemeConfig = defineStore('themeConfig', {
 			 * 全局主题
 			 */
 			// 默认 primary 主题颜色
-			primary: '#DE2910', // '#11559C', // '#03663D'
+			primary: '#409eff',
 			// 是否开启深色模式
 			isIsDark: false,
 
@@ -26,9 +26,9 @@ export const useThemeConfig = defineStore('themeConfig', {
 			 * 顶栏设置
 			 */
 			// 默认顶栏导航背景颜色
-			topBar: '#FFFFFF',
+			topBar: '#ffffff',
 			// 默认顶栏导航字体颜色
-			topBarColor: '#000000',
+			topBarColor: '#606266',
 			// 是否开启顶栏背景颜色渐变
 			isTopBarColorGradual: false,
 
@@ -36,11 +36,11 @@ export const useThemeConfig = defineStore('themeConfig', {
 			 * 菜单设置
 			 */
 			// 默认菜单导航背景颜色
-			menuBar: '#FFFFFF',
+			menuBar: '#545c64',
 			// 默认菜单导航字体颜色
-			menuBarColor: '#000000',
+			menuBarColor: '#eaeaea',
 			// 默认菜单高亮背景色
-			menuBarActiveColor: 'var(--el-color-primary-light-7)',
+			menuBarActiveColor: 'rgba(0, 0, 0, 0.2)',
 			// 是否开启菜单背景颜色渐变
 			isMenuBarColorGradual: false,
 
@@ -48,9 +48,9 @@ export const useThemeConfig = defineStore('themeConfig', {
 			 * 分栏设置
 			 */
 			// 默认分栏菜单背景颜色
-			columnsMenuBar: '#2C3A49',
+			columnsMenuBar: '#545c64',
 			// 默认分栏菜单字体颜色
-			columnsMenuBarColor: '#F0F0F0',
+			columnsMenuBarColor: '#e6e6e6',
 			// 是否开启分栏菜单背景颜色渐变
 			isColumnsMenuBarColorGradual: false,
 			// 是否开启分栏菜单鼠标悬停预加载(预览菜单)
@@ -64,7 +64,7 @@ export const useThemeConfig = defineStore('themeConfig', {
 			// 是否开启菜单手风琴效果
 			isUniqueOpened: true,
 			// 是否开启固定 Header
-			isFixedHeader: true,
+			isFixedHeader: false,
 			// 初始化变量，用于更新菜单 el-scrollbar 的高度，请勿删除
 			isFixedHeaderChange: false,
 			// 是否开启经典布局分割菜单（仅经典布局生效）
@@ -72,13 +72,13 @@ export const useThemeConfig = defineStore('themeConfig', {
 			// 是否开启自动锁屏
 			isLockScreen: false,
 			// 开启自动锁屏倒计时(s/秒)
-			lockScreenTime: 300,
+			lockScreenTime: 30,
 
 			/**
 			 * 界面显示
 			 */
 			// 是否开启侧边栏 Logo
-			isShowLogo: true,
+			isShowLogo: false,
 			// 初始化变量，用于 el-scrollbar 的高度更新，请勿删除
 			isShowLogoChange: false,
 			// 是否开启 Breadcrumb，强制经典、横向布局不显示
@@ -96,24 +96,24 @@ export const useThemeConfig = defineStore('themeConfig', {
 			// 是否开启 TagsView 共用
 			isShareTagsView: false,
 			// 是否开启 Footer 底部版权信息
-			isFooter: true,
+			isFooter: false,
 			// 是否开启灰色模式
 			isGrayscale: false,
 			// 是否开启色弱模式
 			isInvert: false,
 			// 是否开启水印
-			isWatermark: false,
+			isWartermark: true,
 			// 水印文案
-			watermarkText: '星海科技',
+			wartermarkText: 'vue-next-admin',
 
 			/**
 			 * 其它设置
 			 */
 			// Tagsview 风格：可选值"<tags-style-one|tags-style-four|tags-style-five>"，默认 tags-style-five
 			// 定义的值与 `/src/layout/navBars/tagsView/tagsView.vue` 中的 class 同名
-			tagsStyle: 'tags-style-one',
-			// 主页面切换动画: Animate.css
-			animation: 'fadeLeft',
+			tagsStyle: 'tags-style-five',
+			// 主页面切换动画：可选值"<slide-right|slide-left|opacitys>"，默认 slide-right
+			animation: 'slide-right',
 			// 分栏高亮风格：可选值"<columns-round|columns-card>"，默认 columns-round
 			columnsAsideStyle: 'columns-round',
 			// 分栏布局风格：可选值"<columns-horizontal|columns-vertical>"，默认 columns-horizontal
@@ -121,11 +121,11 @@ export const useThemeConfig = defineStore('themeConfig', {
 
 			/**
 			 * 布局切换
-			 * 注意：为了演示，切换布局时，颜色会被还原成默认，代码位置：/@/layout/navBars/breadcrumb/setings.vue
+			 * 注意：为了演示，切换布局时，颜色会被还原成默认，代码位置：/@/layout/navBars/topBar/setings.vue
 			 * 中的 `initSetLayoutChange(设置布局切换，重置主题样式)` 方法
 			 */
 			// 布局切换：可选值"<defaults|classic|transverse|columns>"，默认 defaults
-			layout: 'columns',
+			layout: 'defaults',
 
 			/**
 			 * 后端控制路由
@@ -137,15 +137,15 @@ export const useThemeConfig = defineStore('themeConfig', {
 			 * 全局网站标题 / 副标题
 			 */
 			// 网站主标题（菜单导航、浏览器当前网页标题）
-			globalTitle: '星海科技',
+			globalTitle: 'vue-next-admin',
 			// 网站副标题（登录页顶部文字）
-			globalViceTitle: '星海科技',
+			globalViceTitle: 'vueNextAdmin',
 			// 网站副标题（登录页顶部文字）
-			globalViceTitleMsg: '有可能是.NET最好用的通用权限开发框架',
+			globalViceTitleMsg: '专注、免费、开源、维护、解疑',
 			// 默认初始语言，可选值"<zh-cn|en|zh-tw>"，默认 zh-cn
 			globalI18n: 'zh-cn',
 			// 默认全局组件大小，可选值"<large|'default'|small>"，默认 'large'
-			globalComponentSize: 'small',
+			globalComponentSize: 'large',
 		},
 	}),
 	actions: {